create utilities for Cnalde <-->ndarray conversion
