+++ ferritin-onnx-models/src/lib.rs
@@ -1,5 +1,7 @@
 pub mod models;
+pub mod utilities;
 
 // pub use models::amplify::{AMPLIFYModels, AMPLIFY};
 pub use models::esm2::{ESM2Models, ESM2};
 pub use models::ligandmpnn::{LigandMPNN, LigandMPNNModels};
+pub use utilities::{ndarray_to_tensor_f32, tensor_to_ndarray_f32, tensor_to_ndarray_i64};
